rename main package and root module
