+++ .idea/.name
@@ -1,1 +1,1 @@
-Lecture1
+Сracker
