+++ app/build.gradle
@@ -9,13 +9,11 @@ android {
     compileSdk 31
 
     defaultConfig {
-        applicationId "com.example.lecture1"
+        applicationId "com.example.cracker"
         minSdk 21
         targetSdk 31
         versionCode 1
         versionName "1.0"
-
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
 
     buildTypes {
