+++ app/src/main/AndroidManifest.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.lecture1">
+    package="com.swallow.cracker">
 
     <application
-        android:name=".App"
+        android:name="com.swallow.cracker.App"
         android:allowBackup="false"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -12,7 +12,7 @@
         android:theme="@style/Theme.Lecture1">
 
         <activity
-            android:name=".ui.activitys.MainActivity"
+            android:name="com.swallow.cracker.ui.activitys.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
