+++ app/src/main/java/com/swallow/cracker/utils/MutableLiveData.kt
@@ -1,4 +1,4 @@
-package com.example.lecture1.utils
+package com.swallow.cracker.utils
 
 import androidx.lifecycle.MutableLiveData
