+++ app/src/main/res/values-ru/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">Лекция1</string>
+    <string name="app_name">Сracker</string>
     <string name="icon_smile">значок улыбки</string>
     <string name="skip">Далее</string>
     <string name="welcome_description">Это приветственный экран специально для KTS</string>
