+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">Lecture</string>
+    <string name="app_name">Сracker</string>
     <string name="icon_smile">icon smile</string>
     <string name="skip">Skip</string>
     <string name="welcome_description">This is a welcome screen specially for KTS</string>
