+++ settings.gradle
@@ -3,8 +3,7 @@ dependencyResolutionManagement {
     repositories {
         google()
         mavenCentral()
-        jcenter() // Warning: this repository is going to shut down soon
     }
 }
-rootProject.name = "Lecture1"
+rootProject.name = "Сracker"
 include ':app'
